+++ include/globdef.h
@@ -1,6 +1,8 @@
 #ifndef GLOBDEF_H
 #define GLOBDEF_H
 
+#include <stddef.h>
+
 constexpr unsigned char ERR_UNRECOVERABLE = 1;
 constexpr unsigned char ERR_RECOVERABLE = 0;
 constexpr unsigned short MAX_BUFF_SIZE = 65535;
@@ -12,6 +14,6 @@ constexpr unsigned short MAX_BUFF_SIZE = 65535;
  * @param output The buffer where the extracted token will be stored.
  * @return The length of the json value
  */
-int get_json_value(const char input[], const char key[], char output[]);
+size_t get_json_value(const char input[], const char key[], char output[]);
 
 #endif
