+++ include/window.h
@@ -19,14 +19,14 @@ struct Window {
  * @brief Clears the terminal window.
  * @return An integer status code indicating success or failure.
  */
-int clear_chat_window();
+size_t clear_chat_window();
 
 /**
  * @brief Draws the inner chat window along with context
  * @param window Properties of the window to draw
  * @return Whether the function was successful
  */
-int draw_chat_window(const Window window);
+size_t draw_chat_window(const Window window);
 
 /**
  * @brief Calculate dimensions of the window with respect to context
